Fixed a critical memory access bug...
Added to unsigned long long int data_set_size to the list of parameters for the kernels. Before this, they were throwing a memory access violation bug, because the kernels weren't checking if they were accessing memory that was out of their scope. Also, made both kernels take mean as a double.